third_party/blink/web_tests/platform/win/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-align-expected.txt

This is a testharness.js-based test.
[FAIL] div[style="text-align: center"] legend
  assert_equals: expected legend[align=left] expected 24 but got 395
[FAIL] div[style="text-align: center"][align=center] legend
  assert_equals: expected legend[align=left] expected 24 but got 395
[FAIL] fieldset[dir=rtl] legend[style="text-align: left"]
  assert_equals: expected legend[align=right] expected 765 but got 24
Harness: the test ran to completion.
